Funções randItem e spawFruit movidas para Game.js
